[project @ 2002-10-18 15:26:29 by simonmar] It is better not to use #def in .hsc files, otherwise we have to arrange to include the appropriate _hsc.h file in the package.
There was one use of #def in Network.Socket: my_inet_ntoa, and because
of a bug in the build system the _hsc.o file wasn't being included in
the library when SPLIT_OBJS is on.  The result is that if you compile
against the net package in 5.04.1 and turn on optimisation (so that
inet_ntoa gets inlined), you get a link error.

This commit moves my_inet_ntoa into the HsNet.h header file, and keeps
a static version in the library just in case the compiler fails to
inline it (the same way we do it for all the inline functions in the
base package).